# Implement JSString to reduce bridging overhead