--- a/Sources/JavaScriptKit/Compatibility.swift
+++ b/Sources/JavaScriptKit/Compatibility.swift
@@ -3,5 +3,5 @@
 /// this and `SwiftRuntime.version` in `./Runtime/src/index.ts`.
 @_cdecl("swjs_library_version")
 func _library_version() -> Double {
-    return 610
+    return 611
 }